Calculate invertebrate mass

Calculates invertebrate mass (kg) based on field acreage. Requires input of the fields data frame with area_ac attribute. Returns the data frame with additional columns for calculated values. Can be chained with pipes.

calc_inv_mass(df, day)

Arguments

df

An sf dataset, such as the provided dataframe fields, which includes an attribute area_ac containing the field area in acres. Also works with an ordinary data.frame or tibble.

day

The day number for which to calculate the invertebrate mass. If not provided, then only a daily mass is calculated.

Examples

# calculate the invertebrate mass after 14 days for each field
fields |> calc_inv_mass(14)
#> Simple feature collection with 10106 features and 8 fields
#> Geometry type: POLYGON
#> Dimension:     XYZ
#> Bounding box:  xmin: 6478160 ymin: 1939404 xmax: 6762226 ymax: 2383039
#> z_range:       zmin: 0 zmax: 0
#> Projected CRS: NAD83(2011) / California zone 2 (ftUS)
#> # A tibble: 10,106 × 9
#>    unique_id county                  geometry group_id area_ac volume_af area_m2
#>  * <chr>     <chr>  <POLYGON [US_survey_foot> <chr>      <dbl>     <dbl>   <dbl>
#>  1 1103539   Glenn  Z ((6526707 2289953 0, 6… 1802010…   23.4      9.76   94815.
#>  2 1102849   Glenn  Z ((6517766 2307611 0, 6… 1802010…   23.5      9.78   95009.
#>  3 1103353   Glenn  Z ((6501679 2309168 0, 6… 1802010…    1.06     0.440   4270.
#>  4 1103425   Glenn  Z ((6525986 2324730 0, 6… 1802010…    1.34     0.558   5422.
#>  5 1103546   Glenn  Z ((6535820 2277869 0, 6… 1802010…   48.9     20.4   197899.
#>  6 1103560   Glenn  Z ((6499898 2267008 0, 6… 1802010…   15.1      6.27   60943.
#>  7 1103561   Glenn  Z ((6505757 2270476 0, 6… 1802010…   34.7     14.5   140544.
#>  8 1103563   Glenn  Z ((6544553 2330567 0, 6… 1802010…   24.9     10.4   100575.
#>  9 1103564   Glenn  Z ((6529293 2334151 0, 6… 1802010…   74.7     31.1   302240.
#> 10 1103566   Glenn  Z ((6551712 2337179 0, 6… 1802010…   56.0     23.3   226766.
#> # ℹ 10,096 more rows
#> # ℹ 2 more variables: daily_prod_kg <dbl>, total_prod_kg <dbl>