Calculate invertebrate mass time series

A wrapper for calc_inv_mass that runs the calculation for a series of elapsed days from 1 to ndays. Returns a data frame with total invertebrate mass (grams) by field unique_id and day.

calc_inv_mass_ts(df, ndays)

Arguments

df

An sf dataset, such as the provided dataframe fields, which includes an attribute area_ac containing the field area in acres. Also works with an ordinary data.frame or tibble.

ndays

The total number of days over which to conduct the mass calculation. A sequence of days from 1 to ndays is fed into the day parameter of the calc_inv_mass function.

Examples

# calculate the invertebrate mass time series
fields |> calc_inv_mass_ts(14)
#> # A tibble: 141,484 × 3
#>      day unique_id total_prod_g
#>    <int> <chr>            <dbl>
#>  1     1 1103539      0.00108  
#>  2     1 1102849      0.00108  
#>  3     1 1103353      0.0000485
#>  4     1 1103425      0.0000616
#>  5     1 1103546      0.00225  
#>  6     1 1103560      0.000692 
#>  7     1 1103561      0.00160  
#>  8     1 1103563      0.00114  
#>  9     1 1103564      0.00343  
#> 10     1 1103566      0.00258  
#> # ℹ 141,474 more rows

# calculate the invertebrate mass time series for a particular field
fields |> dplyr::filter(unique_id=="1103539") |> calc_inv_mass_ts(14)
#> # A tibble: 14 × 3
#>      day unique_id total_prod_g
#>    <int> <chr>            <dbl>
#>  1     1 1103539        0.00108
#>  2     2 1103539        0.00215
#>  3     3 1103539        0.00323
#>  4     4 1103539        0.00431
#>  5     5 1103539        0.00538
#>  6     6 1103539        0.00646
#>  7     7 1103539        0.00754
#>  8     8 1103539        0.00861
#>  9     9 1103539        0.00969
#> 10    10 1103539        0.0108 
#> 11    11 1103539        0.0118 
#> 12    12 1103539        0.0129 
#> 13    13 1103539        0.0140 
#> 14    14 1103539        0.0151