Create leaflet map object

Initialize a leaflet map object with the desired basemap and settings.

ff_make_leaflet(
  bbox = c(xmin = -122.3, ymin = 38.5, xmax = -121.3, ymax = 39.7)
)

Arguments

bbox

A named vector defining the initial map extent via minimum and maximum lat/lon coordinates in the format c(xmin=_, ymin=_, xmax=_, ymax=_)

Examples

# create and display leaflet object with a specified bounding box
bbox <- sf::st_bbox(ff_fields_gcs)
ff_make_leaflet(bbox)

# create the leaflet map for a Shiny app
if(FALSE){
  output$mainMap <- leaflet::renderLeaflet({
    ff_make_leaflet(bbox)
  })
}